pyproject.toml
[tool.poetry]
name = "рщhw"
version = "0.1.0"
description = ""
authors = ["ArtemSermyagin <artem.sermyagin@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
